requirements.txt
weaviate-client[agents]>=4.0.0
datasets>=2.14.0
python-dotenv>=1.0.0 